feat: add test for coverage
